separate out deprecated functions by build constraints
